requirements/base.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
idna==3.11
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
pandas==2.3.3
pca==2.10.1
djangorestframework==3.16.1
pydantic==2.12.5
